+++ NEWS.md
@@ -1,3 +1,9 @@
+# calcofi4r 0.3.0
+
+* + `get_cruises()` and `get_raster()` for reading from the CalCOFI API.
+* + `map_raster()` and `plot_timeseries()` for visualizing.
+* + "Get started" vignette and README.Rmd -> README.md with example usage
+
 # calcofi4r 0.2.0
 
 * + `get_variables()`, `get_timeseries()` for reading from the API at https://api.calcofi.io